generate .pdf files with ordered data

+++ controllers/admin.js
@@ -1,5 +1,7 @@
 const mongoose = require('mongoose')
 
+const fileHelper = require('../util/file');
+
 const {
     validationResult
 } = require('express-validator/check');
@@ -151,6 +153,7 @@ exports.postEditProduct = (req, res, next) => {
             }
             product.title = updatedTitle;
             if (image) {
+                fileHelper.deleteFile(product.imageUrl);
                 product.imageUrl = image.path;
             }
             product.price = updatedPrice;
@@ -190,9 +193,16 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
     const prodId = req.body.productId;
-    Product.deleteOne({
-            _id: prodId,
-            userId: req.user._id
+    Product.findById(prodId)
+        .then(product => {
+            if (!product) {
+                return next(new Error('product not found'));
+            }
+            fileHelper.deleteFile(product.imageUrl);
+            return Product.deleteOne({
+                _id: prodId,
+                userId: req.user._id
+            })
         })
         .then(() => {
             console.log('product deleted')
